backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
alembic==1.14.1
python-dotenv==1.0.1
httpx>=0.27.0
lancedb>=0.27.0
pyyaml>=6.0
